Created image input componment
